refactor(inline): update slots and extract text service
