+++ packages/inline/src/utils/index.ts
@@ -1,6 +1,7 @@
 export * from './attribute-renderer.js';
 export * from './base-attributes.js';
 export * from './delta-convert.js';
+export * from './embed.js';
 export * from './guard.js';
 export * from './keyboard.js';
 export * from './point-conversion.js';
