+++ packages/inline/src/utils/range-conversion.ts
@@ -2,7 +2,7 @@ import type * as Y from 'yjs';
 
 import { VElement } from '../components/v-element.js';
 import type { InlineRange } from '../types.js';
-import { isInEmbedElement } from './guard.js';
+import { isInEmbedElement } from './embed.js';
 import {
   nativePointToTextPoint,
   textPointToDomPoint,
